# public/openapi.yaml
openapi: 3.1.0
info:
  title: BuildingSync API
  version: "0.1.0"
  summary: >-
    Public HTTP API for the BuildingSync platform. Designed to be consumed by
    the BuildingSync web app, native iOS and Android apps, and authorized
    third-party integrations.
  description: |
    BuildingSync is the operating system for residential buildings. This
    API lets clients submit and track maintenance work orders, broadcast
    announcements, manage Web Push subscriptions, and export personal data.

    ## Authentication

    Two authentication modes are supported on the same endpoints:

    - **Cookie session** (web). After signing in through the web app, the
      Supabase session cookie is forwarded automatically by the browser.
    - **Bearer JWT** (mobile / native). Native clients sign in directly via
      the Supabase Auth SDK and pass the resulting access token as
      `Authorization: Bearer <access_token>` on every request. Tokens
      should be refreshed using the Supabase SDK's standard refresh flow.

    Both modes resolve to the same user record. New clients should prefer
    Bearer authentication.

    ## Real-time updates

    For low-latency sync, mobile clients should subscribe to row-level
    Postgres changes via Supabase Realtime using the same Supabase project
    credentials. This keeps work-order status, announcements, and
    deliveries in sync across devices without polling. Push notifications
...

    ## Versioning

...

  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
  - cookieSession: []
paths:
...
  /api/me/building:
    get:
      tags: [Me]
      summary: Everything the signed-in user can see for their building.
      description: |
        Aggregate read for the resident/tenant mobile surface: building info,
        announcements, deliveries addressed to the user, the user's maintenance
        requests, payment status (tenants), and move-out checklist (tenants).
        One round-trip keeps the mobile launch fast; clients may also call the
        per-resource endpoints below for refresh.
      responses:
        "200":
          description: Building data bundle.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BuildingData" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/me/maintenance:
    get:
      tags: [Me]
      summary: The caller's maintenance requests.
      responses:
        "200":
          description: Maintenance requests.
          content:
            application/json:
              schema:
                type: object
                required: [requests]
                properties:
                  requests:
                    type: array
                    items: { $ref: "#/components/schemas/MaintenanceRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Me]
      summary: Submit a new maintenance request.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, area]
              properties:
                title: { type: string, minLength: 3, maxLength: 200 }
                area: { type: string, maxLength: 80 }
                description: { type: string, maxLength: 2000 }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MaintenanceRequest" }
        "400": { $ref: "#/components/responses/InvalidBody" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Access token obtained from POST /api/auth/login (or /api/auth/refresh).
        Backend-agnostic: the token is issued by whatever auth provider backs
        the platform (currently Supabase; Auth.js when BuildingSync moves to
        Neon). Clients never depend on the provider — only on this contract.
    cookieSession:
      type: apiKey
      in: cookie
      name: bs-session
      description: HttpOnly session cookie set by the web app after sign-in.
  responses:
    Unauthorized:
      description: The caller is not signed in.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InvalidBody:
      description: The request body did not match the expected schema.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: unauthorized
        message:
          type: string
          example: Your account is not yet linked to a building.
        issues:
          type: array
          description: Optional Zod-style validation issue list.
          items: { type: object }
...
    Announcement:
      type: object
      required: [id, buildingId, title, body, audience, createdAt]
      properties:
        id: { type: string, format: uuid }
        buildingId: { type: string, format: uuid }
        authorId: { type: string, format: uuid }
        title: { type: string }
        body: { type: string }
        audience:
          type: string
          enum: [all, tenants_only, specific_units]
        targetUnitIds:
          type: array
          items: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
        deletedAt: { type: string, format: date-time, nullable: true }
...
    Building:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Building One }
        address: { type: string, nullable: true }
        city: { type: string, nullable: true }
        registry: { type: string, nullable: true, description: Condo/legal registry id, e.g. TSCC 1694. }
    Delivery:
      type: object
      required: [id, status]
      properties:
        id: { type: string, format: uuid }
        from: { type: string, nullable: true, example: Amazon }
        description: { type: string, nullable: true }
        arrivedAt: { type: string, format: date-time, nullable: true }
        code: { type: string, nullable: true, example: B-2244 }
        status: { type: string, enum: [ready, picked_up] }
    MaintenanceRequest:
      type: object
      required: [id, title, status]
      properties:
        id: { type: string }
        title: { type: string }
        area: { type: string, nullable: true }
        status: { type: string, enum: [open, in_progress, scheduled, resolved] }
        submittedAt: { type: string, format: date-time, nullable: true }
        lastUpdate: { type: string, nullable: true }
    PaymentInfo:
      type: object
      properties:
        amountDue: { type: number, example: 2850 }
        currency: { type: string, example: CAD }
        dueDate: { type: string, format: date, nullable: true }
        daysUntilDue: { type: integer, nullable: true }
        autopay: { type: boolean }
    BuildingData:
      type: object
      required: [building, announcements, deliveries, maintenance]
      description: Aggregate read for the resident/tenant mobile home surface.
      properties:
        building: { $ref: "#/components/schemas/Building" }
        announcements:
          type: array
          items: { $ref: "#/components/schemas/Announcement" }
        deliveries:
          type: array
          items: { $ref: "#/components/schemas/Delivery" }
        maintenance:
          type: array
          items: { $ref: "#/components/schemas/MaintenanceRequest" }
        payment:
          oneOf:
            - { $ref: "#/components/schemas/PaymentInfo" }
            - { type: "null" }
          description: Present for tenants; null otherwise.